refactor assets and events, spruce up error handling and fix issues with uframe not loading
