+++ ooiservices/app/uframe/__init__.py
@@ -10,5 +10,5 @@ from flask import Blueprint
 
 uframe = Blueprint('uframe', __name__)
 
-from ooiservices.app.uframe import controller, assetController,data
+from ooiservices.app.uframe import controller, assetController, data, assets, events
